requirements.txt
pip
python=3.10
torch
torchvision
pandas
pillow
matplotlib